+++ docs/changelog.rst
@@ -6,7 +6,13 @@ CHANGELOG
 ----------------------------
 
 **Improvements**
+
 - Change provider Charfield for a foreign key
+
+**Bug fixes**
+
+- Fix parser crash after exception during attachment download (refs #4807)
+
 
 
 2.116.0    (2025-07-03)